requirements.txt
flask
mysql-connector-python
cryptography
pymysql